# Reduce overhead accessing process.env in benchmark