@divmain 2017-03-02
Thank you!  It's a downer I didn't add this before launch, but I'd much rather be providing accurate perf data.  My benchmarks don't match up with yours too well - what type of machine is this running on?