@divmain 2017-03-02
Here's the current benchmarks on my machine, for reference:
```
Starting benchmark for 8 concurrent render operations...
renderToString took 8.630521409 seconds
rapscallion, no caching took 9.591807761 seconds; ~0.89x faster
rapscallion, caching DIVs took 3.348522824 seconds; ~2.57x faster
rapscallion, caching DIVs (second time) took 2.392897817 seconds; ~3.6x faster
rapscallion, caching Components took 2.202549012 seconds; ~3.91x faster
rapscallion, caching Components (second time) took 2.058421309 seconds; ~4.19x faster
```